fix: move button on user list to top right
