+++ src/page/admin/users/UsersList/index.js
@@ -1,9 +1,11 @@
 import { connect } from 'react-redux';
 import UsersList from './UsersList';
 
-const mapStateToProps = state => ({
-    location: state.settings.toJS().location || {},
-});
+const mapStateToProps = state => {
+    return {
+        location: state.settings.toJS().location || {},
+    };
+};
 
 const Container = connect(mapStateToProps)(UsersList);
